fix: rename to wsl filename
